Update job details page (#127)
* Show values in input/output menu

Display input/output values since not
all of the values will be links for cromwell
workflows. For values that are links to files,
display just the file name.

* Fix url parsing error

Check whether the url param
in validateGcsURLGetParts is a
string before trying to split
the string.

* Override default menu width

To prevent the values in the inputs and
outputs menus on the job details page from
getting cut off, set the max-width of those
menus to none. Also handle long values by
truncating any lines that are > 100 characters.

* Convert TaskMetadata dates

Convert start and end times of each
task into datetime objects in addition
to converting the job times.

* Update information displayed for tasks

* Consolidate duplicated formatting functions

The task and panel components both require the
same functions for formatting links and getting
the duration of a job or task. Move these functions
into a separate class

* Add labels to job details page

* Add pipe for job/task duration

* Link directly to task log files

* Create separate file for ResourceUtils

* Update ng lint rules

Remove no-access-missing-member rule since
the linter incorrectly throws this error for
methods that a class inherits, and this error
will already get caught by Angular's compiler

* Update html

* Fix panel component tests

* Add tests for tasks component

* Add tests for duration pipe

* Add resource-utils class with tests

* Fix input/outputs menu width

Instead of shortening the length of
the text to prevent long values from
thowing off the UI, change the max-width of
the menu

* Minor formatting & naming changes

* Make ResourceUtils methods static

* Change label format

* Formatting changes, code clean-up

* Fix usage of ResourceUtils methods