List cromwell jobs newest first
Use the totalResultsCount from the
query endpoint to get the jobs starting
from the last page of results since they
are returned oldest first in cromwell.

Once the total number of jobs for the query
is known, fetch the page from the end of the
results based on the current offset and continue
fetching jobs if the number is less than the min
number to display.